refactor: replace createCartItem to class CartItem
